LayoutTests/imported/w3c/web-platform-tests/html/links/stylesheet/quirk-origin-check-expected.txt

CONSOLE MESSAGE: Did not parse stylesheet at 'data:text/plain,.test { background: red }' because non CSS MIME types are not allowed for cross-origin stylesheets.
There should be no red.


PASS Origin check for stylesheet with non-CSS MIME type quirk
